src/API/safe/requirements.txt
fastapi[standard]
pydantic
pydantic-settings
pytest
pylint
sqlalchemy[asyncio]
asyncpg
logging
argon2-cffi